Generate a meme

Generate a meme with a background image, text label and optional plot.

meme(img, label, file, g, size = 7, fontfamily = "Impact", col = "white",
  shadow = "black", width, height, mult = 1, g_pos = list(width = 0.9,
  height = 0.6, x = 0.5, y = 0.35), label_pos = list(width = 0.9, height =
  0.3, x = 0.5, y = 0.9), ggtheme, panel_background = "#FFFFFF50",
  plot_background = "#FFFFFF50")

Arguments

img
path to image file, .png or .jpg.

label
character, meme text. May be a vector, matched to lab_pos.

file
output file, .png or .jpg.

g
a ggplot object. This is an optional inset plot and may be excluded.

size
label size.

fontfamily
character, defaults to "Impact", the classic meme font.

col
label color.

shadow
label shadow/outline color.

width
numeric, width of overall meme plot in pixels. If missing, taken from img size.

height
numeric, height of overall meme plot in pixels. If missing, taken from img size.

mult
numeric, a multiplier. Used to adjust width and height. See details.

g_pos
list of position elements for the g inset plot.

label_pos
list of position elements for the meme text. Each element may be a vector. See details.

ggtheme
optional ggplot2 theme. If ignored, the default memery ggplot2 theme is used.

panel_background
color, overrides ggplot theme. See details.

plot_background
color, overrides ggplot theme. See details.

Details

This function generates a meme and saves to disk as a png. The meme plot may optionally include an inset plot by passing a ggplot object to g. This makes the memes more fun for data analysts. See examples.

List elements in lab_pos must all be the same length and must match the length of label. This is provided for generality but is most suited to length-2 cases; the use of meme title/subtitle or top/bottom text pairs. mult is typically set less than one if relying on img dimension for meme plot width and height and img is large. panel_background ans plot_bakcground should be transparent or semi-transparent if occupying a large amount of the meme plot or it will obscure the meme image. An alternative is to use g as a tiny thumbnail in the corner of the meme plot, in which case full opacity is not an issue.

Examples

# NOT RUN {
meme("image.png", ggObject, "My first memery meme!", "myfirstmeme.png")
# }