Meme plot inset templates

Templates for the position and background of a ggplot inset.

inset_position(type = "default")

inset_background(type = "default")

inset_templates(type)

Arguments

type
character, name of template.

Value

a list of arguments passed to either inset_pos or inset_background in meme.

Details

inset_position and inset_background assist with some basic options for position and background of the optional ggplot inset graphic. inset_templates can be used to view the available templates. See examples. If a template is not available to suit your needs, provide your own argument list to meme.

Examples

inset_templates("position")

#> [1] "default" "topleft"

inset_templates("background")

#> [1] "default"

inset_position()

#> $w
#> [1] 0.95
#> 
#> $h
#> [1] 0.6
#> 
#> $x
#> [1] 0.5
#> 
#> $y
#> [1] 0.325
#> 

inset_background()

#> $fill
#> [1] "#FFFFFF50"
#> 
#> $col
#> [1] NA
#> 
#> $r
#> [1] 0.025snpc
#>